requirements.txt
streamlit
pillow
opencv-contrib-python
ultralytics
torch